% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barriers}
\alias{barriers}
\title{Fish Passage Barriers (2025)}
\format{
A tibble with 38 rows and 9 variables.
\itemize{
  \item \code{location}: mainstem location
  \item \code{stream_name}: stream or river name
  \item \code{barrier_type}: type of passage barrier
.  \item \code{barrier_status}: status of passage barrier (partial, blocked)
  \item \code{source_entity}: entity that provides the data
  \item \code{species}: species impacted
  \item \code{latitude}: latitude of data location
  \item \code{longitude}: longitude of data location
  \item \code{geometry}: the spatial geometry associated with each extent
  }
}
\usage{
barriers
}
\description{
A spatial dataset of priority fish barriers in the Klamath Basin. Data source: https://nrimp.dfw.state.or.us/DataClearinghouse/default.aspx?p=202&XMLname=1094.xml
}
\keyword{datasets}
