% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_extents}
\alias{habitat_extents}
\title{Salmon Habitat Data}
\format{
A tibble with 744 rows and 11 variables.
\itemize{
  \item \code{location}: stream or river name
  \item \code{sub_basin}: sub-basin name (upper klamath, lower klamath, trinity)
  \item \code{data_type}: data type = fish habitat extents
  \item \code{species}: species name (coho, steelhead, chinook)
  \item \code{lifestage}: fish life stage
  \item \code{run}: species run (summer, winter, mixed, unknown, spring, fall, NA)
  \item \code{extent}: pre-or-post dam removal extent
  \item \code{geometry}: the spatial geometry associated with each extent
  }
}
\usage{
habitat_extents
}
\description{
A spatial dataset of linear stream segments in the Klamath Basin that define the extent of salmonid habitat. These features represent:
\itemize{
  \item The reaches or stream sections for which abundance data apply (e.g., survey areas)
  \item Known or modeled distribution areas for steelhead, coho, and Chinook salmon
  \item Areas where these species are currently present or have been extirpated
}
}
\keyword{datasets}
