% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitat_extents}
\alias{habitat_extents}
\title{Salmon Habitat Data}
\format{
A tibble with 299 rows and 11 variables.
\itemize{
  \item \code{stream}: stream
  \item \code{sub_basin}: sub-basin name (upper klamath, lower klamath, trinity)
  \item \code{data_type}:
  \item \code{location_name}: location name
  \item \code{species}:
  \item \code{species_full_name}:
  \item \code{lifestage}:
  \item \code{run}:
  \item \code{latest_year}:
  \item \code{link}: web link containing more information about data
  \item \code{latitude}: latitude of data location
  \item \code{longitude}: longitude of data location
  }
}
\usage{
habitat_extents
}
\description{
A spatial dataset of linear stream segments in the Klamath Basin that define the extent of salmonid habitat. These features represent:
\itemize{
  \item The reaches or stream sections for which abundance data apply (e.g., survey areas)
  \item Known or modeled distribution areas for steelhead, coho, and Chinook salmon
  \item Areas where these species are currently present or have been extirpated
}
}
\keyword{datasets}
